requirements.txt
h5py==2.8.0
scipy==1.1.0
matplotlib==2.2.2
seaborn==0.8.1
click==6.7
pytest==3.6.3
PyYAML==3.13
tqdm==4.32.1
